Restructured complete project
