+++ .gitignore
@@ -3,3 +3,5 @@
 .pytest_cache
 .ipynb_checkpoints
 workshop.ipynb
+.DS_Store
+*.png
