+++ pages/__init__.py
@@ -1,25 +1,5 @@
-from .CreateComplaint import AddComplaintPage
-from .CreateComplaint import ComplaintTypePage
-
-from .Navigation import TopMenuPage
-from .Navigation import BottomMenuPage
-
-from .LanguageSelectionPage import LanguageSelectionPage
-
-from .ComplaintFeedbackPage import ComplaintFeedbackPage
-from .ComplaintUnassignPage import ComplaintUnassignPage
-from .ComplaintReassignPage import ComplaintReassignPage
-
-from .HomePage import HomePage
-
-from .Registration import OTPPage
-
-from .Login import LoginPage
-
-from .Profile import ProfilePage
-
-from .ReopenComplaint import ReopenComplaintPage
-
-from .ComplaintSubmitted import ComplaintSubmittedPage
-
-from .UserRegistration import RegistrationPage
+from .citizen import *
+from .app import *
+from .components import *
+from .employee import *
+from .wip import *
