+++ pages/citizen/__init__.py
@@ -1,0 +1,4 @@
+from .complains import *
+from .common import *
+from .registration import *
+from .profile import *
