+++ tests/test.py
@@ -1,19 +1,7 @@
 from pytest import fixture
 
-from pages import RegistrationPage
-from pages import OTPPage
-from pages import HomePage
-from pages import LanguageSelectionPage
-from pages import AddComplaintPage
-from pages import ComplaintFeedbackPage
-from pages import LoginPage
-from pages import ComplaintUnassignPage
-from pages import ComplaintReassignPage
-
+from pages import *
 from framework.selenium_plus import *
-from pages import CitizenProfilePage
-from pages import ComplaintSubmittedPage
-from pages import MyComplaintsPage
 
 
 # def pytest_sessionstart(session):
@@ -21,7 +9,6 @@ from pages import MyComplaintsPage
 #
 # def pytest_sessionfinish(session, exitstatus):
 #     # teardown_stuff
-
 
 @fixture(autouse=True, scope='session')
 def my_fixture():
@@ -34,50 +21,44 @@ def my_fixture():
         pass
     # teardown_stuff
 
-from pages import OTPPage
-from pages import AddComplaintPage
-from pages import ComplaintTypePage
-from pages import TopMenuPage
-from pages import BottomMenuPage
-from pages import ProfilePage
-from pages import ReopenComplaintPage
-from pages import LoginPage
-
 
 def test_otp_submission():
-    otp = OTPPage
+    otp = OTPPage()
     otp.navigate()
 
-    otp.set("1234").submit()
+    otp.set("12345").get_started()
     assert get_url() == "http://egov-micro-dev.egovernments.org/app/v3/citizen"
 
 
 def test_create_complaint():
-    comp = AddComplaintPage
+    comp = AddComplaintPage()
     comp.navigate()
+    comp.set_complaint_type("Overflowing Garbage Bins")
+    comp.set_location_by_address("Homigo Ant")
     photo1 = "/Users/tarun.lalwani/Documents/screenshots/CapturFiles_1.png"
     photo2 = "/Users/tarun.lalwani/Documents/screenshots/CapturFiles-20180126_012014.png"
     photo3 = "/Users/tarun.lalwani/Documents/screenshots/CapturFiles-20180310_095257.png"
     comp.upload_images(photo1, photo2, photo3)
     comp.remove_image_2()
-    # comp.set_complaint_type("garbage", "Overflowing Garbage Bins")
-    # comp.set_location_by_address("Homigo Ant")
     assert get_url() == "http://egov-micro-dev.egovernments.org/app/v3/citizen/add-complaint"
 
 
 def test_citizen_profile():
-    CitizenProfilePage.navigate().set("ABH", "ab.se@gmail.com").set_city("Amritsar").click_continue()
+    cp = CitizenProfilePage()
+
+    cp.navigate().set("ABH", "ab.se@gmail.com").set_city("Amritsar").save()
 
 
 def test_complain_submitted():
-    ComplaintSubmittedPage.navigate()
-    print(ComplaintSubmittedPage.get_complaint_number())
-    ComplaintSubmittedPage.click_continue()
+    complaint_subpage = ComplaintSubmittedPage()
+    complaint_subpage.navigate()
+    print(complaint_subpage.get_complaint_number())
+    complaint_subpage.click_continue()
     assert get_url() == "http://egov-micro-dev.egovernments.org/app/v3/citizen"
 
 
 def test_my_complaints():
-    complaints = MyComplaintsPage
+    complaints = MyComplaintsPage()
     complaints.navigate()
     cards = complaints.get_all_complaints()
     card = cards[2]
@@ -91,14 +72,14 @@ def test_my_complaints():
 
 
 def test_user_registration():
-    userRegistration = RegistrationPage
-    userRegistration.navigate()
-    userRegistration.set("9988776655", "FirstName", "Bathi").submit()
+    user_reg = RegistrationPage()
+    user_reg.navigate()
+    user_reg.set("9988776655", "FirstName", "Bathinda").submit()
     assert get_url() == "http://egov-micro-dev.egovernments.org/app/v3/citizen/user/otp"
 
 
 def test_language_selection():
-    ls = LanguageSelectionPage
+    ls = LanguageSelectionPage()
     ls.navigate()
     ls.language("punjabi").language("hindi").language("english").submit()
 
@@ -106,7 +87,7 @@ def test_language_selection():
 
 
 def test_homepage():
-    hp = HomePage
+    hp = HomePage()
 
     hp.navigate().new_complaint()
     assert get_url() == "http://egov-micro-dev.egovernments.org/app/v3/citizen/add-complaint"
@@ -116,7 +97,7 @@ def test_homepage():
 
 
 def test_complaintfeedbackpage():
-    cf = ComplaintFeedbackPage
+    cf = ComplaintFeedbackPage()
     cf.navigate().star_click(1)
     cf.check_services().check_quality_of_work()
     cf.set("good to go").submit()
@@ -124,53 +105,50 @@ def test_complaintfeedbackpage():
 
 
 def test_complaintunassignedpage():
-    cap=ComplaintUnassignPage
+    cap = ComplaintUnassignPage()
     cap.navigate().assign().reject()
 
     assert get_url() == "http://egov-micro-dev.egovernments.org/app/stv3/citizen/add-complaint"
 
 
 def test_login():
-    LoginPage.navigate().set("1234567890").submit()
+    LoginPage().navigate().set("1234567890").submit()
+    OTPPage().set("12345").get_started()
     assert get_url() == "http://egov-micro-dev.egovernments.org/app/v3/citizen"
 
 
 def test_reopen_complaint():
-    ReopenComplaintPage.navigate().set("Complaint not resolved").submit()
+    ReopenComplaintPage().navigate().set("Complaint not resolved").submit()
     assert get_url() == "http://egov-micro-dev.egovernments.org/app/v3/citizen/complaint-submitted"
-
-
-def test_complaint_details():
-    ComplaintTypePage.navigate()
-    assert get_url() == "http://egov-micro-dev.egovernments.org/app/v3/citizen/complaint-details?status=filed"
 
 
 def test_navigation():
     LoginPage.navigate().set("9999999999").submit()
     assert get_url() == "http://egov-micro-dev.egovernments.org/app/v3/citizen"
-    BottomMenuPage.info().payments().complaints().home()
-    TopMenuPage.ham()
-    LoginPage.profile()
+    BottomMenuComponent().info().payments().complaints().home()
+    TopMenuNavigationComponent().ham()
+    LoginPage().profile()
     assert get_url() == "http://egov-micro-dev.egovernments.org/app/v3/citizen/user/profile"
-    TopMenuPage.back()
+    TopMenuNavigationComponent().back()
     assert get_url() == "http://egov-micro-dev.egovernments.org/app/v3/citizen"
 
     # this test is for verifying navigation
 
 
 def test_profile():
-    LoginPage.navigate().set("9999999999").submit()
+    LoginPage().navigate().set("9999999999").submit()
+    OTPPage().set("12345").get_started()
     assert get_url() == "http://egov-micro-dev.egovernments.org/app/v3/citizen"
-    TopMenuPage.ham()
-    LoginPage.profile()
+    TopMenuNavigationComponent().ham()
+    LoginPage().profile()
     assert get_url() == "http://egov-micro-dev.egovernments.org/app/v3/citizen/user/profile"
-    ProfilePage.update("Singh", "def@ulb.in")
-    ProfilePage.photo_remover()
-    ProfilePage.save()
-    TopMenuPage.back()
+    ProfilePage().update("Singh", "def@ulb.in")
+    ProfilePage().photo_remove()
+    ProfilePage().save()
+    TopMenuNavigationComponent().back()
     assert get_url() == "http://egov-micro-dev.egovernments.org/app/v3/citizen"
 
-def test_complaintreassignpage():
-    crp=ComplaintReassignPage
-    crp.navigate().reject().assign()
 
+# def test_complaintreassignpage():
+#     crp = ComplaintReassignPage()
+#     crp.navigate().reject().assign()
